updated wifi driver

+++ COTS/Inc/HAL/WiFi/WiFi_private.h
@@ -10,6 +10,6 @@
 
 void WiFi_voidEnter(WiFi_t* module);
 
-b8 WiFi_b8ValidateJustExecuted(WiFi_t* module, u16 timeout);
+b8 WiFi_b8ValidateJustExecuted(WiFi_t* module, u32 timeout);
 
 #endif /* INCLUDE_HAL_WIFI_WIFI_PRIVATE_H_ */
